pyproject.toml
[project]
name = "feedee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django>=5.2,<6.0",
  "djangorestframework>=3.17,<4.0",
  "bleach>=6.1",
]
